# organon/backend/openapi.yaml
swagger: "2.0"
info:
  description: "Organon Serice Front"
  title: "Service Front"
  version: "1.0.0"
host: "organon-api.endpoints.organon-srv.cloud.goog"
basePath: "/"
consumes:
- "application/json"
produces:
- "application/json"
schemes:
- "https"
paths:
...
  "/stories/{id}/{subscribe}":
    post:
      description: "Subscribe"
      operationId: "subscription"
      parameters:
      - name: "id"
        in: path
        type: string
        required: true
      - name: "subscribe"
        in: path
        type: string
        required: true
      produces:
      - "application/json"
      responses:
        200:
          description: "Echo"
          schema:
            $ref: "#/definitions/echoMessage"
      x-security:
      - firebase:
          audiences:
          - "organon-srv"
...
definitions:
  echoMessage:
    properties:
      message:
        type: "string"
...
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "key"
    in: "query"
  firebase:
    authorizationUrl: ""
    flow: "implicit"
    type: "oauth2"
    x-google-issuer: "https://securetoken.google.com/organon-srv"
    x-google-jwks_uri: "https://www.googleapis.com/service_accounts/v1/metadata/x509/securetoken@system.gserviceaccount.com"